feat: enhance guitar state measurement with navigation and manual detection; refactor frequency detection and state management

+++ lib/provider/frequency_stream_provider.dart
@@ -1,10 +1,8 @@
-import 'dart:typed_data';
+import 'dart:async';
 
 import 'package:floyd_rose_tuner/provider/audio_stream_provider.dart';
 import 'package:floyd_rose_tuner/utils/convert_pcm16_to_float.dart';
 import 'package:pitch_detector_dart/pitch_detector.dart';
-import 'dart:async';
-
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'frequency_stream_provider.g.dart';
 
@@ -15,8 +13,8 @@ const int bitsPerSample = 8;
 const int bitrate = bitsPerSample*sampleRate;
 
 final pitchDetectorDart = PitchDetector(
-    audioSampleRate: sampleRate*1,
-    bufferSize: PitchDetector.DEFAULT_BUFFER_SIZE as int,
+    audioSampleRate: sampleRate * 1,
+    bufferSize: PitchDetector.DEFAULT_BUFFER_SIZE,
 );
